refactor(file): move file.types to global
