+++ src/global/shared-types/file.types.ts
